feat(#99): move some classes
