+++ src/main/java/com/github/lombrozo/jsmith/antlr/Unparser.java
@@ -24,7 +24,7 @@
 package com.github.lombrozo.jsmith.antlr;
 
 import com.github.lombrozo.jsmith.antlr.rules.Rule;
-import com.github.lombrozo.jsmith.antlr.view.Text;
+import com.github.lombrozo.jsmith.antlr.view.Snippet;
 import java.util.HashMap;
 import java.util.Map;
