+++ src/main/java/com/github/lombrozo/jsmith/antlr/rules/Cage.java
@@ -24,10 +24,8 @@
 package com.github.lombrozo.jsmith.antlr.rules;
 
 import com.github.lombrozo.jsmith.antlr.Context;
-import com.github.lombrozo.jsmith.antlr.LeafSnippet;
-import com.github.lombrozo.jsmith.antlr.Snippet;
-import com.github.lombrozo.jsmith.antlr.view.Text;
-import com.github.lombrozo.jsmith.antlr.view.TextLeaf;
+import com.github.lombrozo.jsmith.antlr.view.LeafSnippet;
+import com.github.lombrozo.jsmith.antlr.view.Snippet;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicReference;
