+++ src/main/java/com/github/lombrozo/jsmith/antlr/rules/Rule.java
@@ -24,8 +24,7 @@
 package com.github.lombrozo.jsmith.antlr.rules;
 
 import com.github.lombrozo.jsmith.antlr.Context;
-import com.github.lombrozo.jsmith.antlr.Snippet;
-import com.github.lombrozo.jsmith.antlr.view.Text;
+import com.github.lombrozo.jsmith.antlr.view.Snippet;
 
 /**
  * Rule definition.
