+++ src/main/java/com/github/lombrozo/jsmith/antlr/rules/Ruleref.java
@@ -24,9 +24,8 @@
 package com.github.lombrozo.jsmith.antlr.rules;
 
 import com.github.lombrozo.jsmith.antlr.Context;
-import com.github.lombrozo.jsmith.antlr.Snippet;
+import com.github.lombrozo.jsmith.antlr.view.Snippet;
 import com.github.lombrozo.jsmith.antlr.Unparser;
-import com.github.lombrozo.jsmith.antlr.view.Text;
 
 /**
  * Parser rule reference.
