+++ src/main/java/com/github/lombrozo/jsmith/antlr/semantic/VariableDeclaration.java
@@ -25,12 +25,11 @@ package com.github.lombrozo.jsmith.antlr.semantic;
 
 import com.github.lombrozo.jsmith.antlr.Attributes;
 import com.github.lombrozo.jsmith.antlr.Context;
-import com.github.lombrozo.jsmith.antlr.LeafSnippet;
-import com.github.lombrozo.jsmith.antlr.Snippet;
+import com.github.lombrozo.jsmith.antlr.view.LeafSnippet;
+import com.github.lombrozo.jsmith.antlr.view.Snippet;
 import com.github.lombrozo.jsmith.antlr.rules.Rule;
 import com.github.lombrozo.jsmith.antlr.view.TextLeaf;
 import java.util.Collections;
-import org.checkerframework.checker.units.qual.A;
 
 /**
  * Variable Declaration Semantic.
