+++ src/main/java/com/github/lombrozo/jsmith/antlr/semantic/VariableTarget.java
@@ -25,13 +25,12 @@ package com.github.lombrozo.jsmith.antlr.semantic;
 
 import com.github.lombrozo.jsmith.antlr.Attributes;
 import com.github.lombrozo.jsmith.antlr.Context;
-import com.github.lombrozo.jsmith.antlr.ErrorSnippet;
-import com.github.lombrozo.jsmith.antlr.LeafSnippet;
-import com.github.lombrozo.jsmith.antlr.Snippet;
+import com.github.lombrozo.jsmith.antlr.view.ErrorSnippet;
+import com.github.lombrozo.jsmith.antlr.view.LeafSnippet;
+import com.github.lombrozo.jsmith.antlr.view.Snippet;
 import com.github.lombrozo.jsmith.antlr.rules.Rule;
 import com.github.lombrozo.jsmith.antlr.view.TextLeaf;
 import com.jcabi.log.Logger;
-import java.util.Collections;
 import java.util.Optional;
 
 /**
